+++ run/README.md
@@ -12,6 +12,7 @@ This directory contains samples for [Google Cloud Run](https://cloud.run). [Clou
 |           Sample                |        Description       |     Deploy    |
 | ------------------------------- | ------------------------ | ------------- |
 |[Hello World][helloworld]&nbsp;&#10149; | Quickstart | [<img src="https://storage.googleapis.com/cloudrun/button.svg" alt="Run on Google Cloud" height="30">][run_button_helloworld] |
+|[Events – Pub/Sub][events_pubsub]| Use Pub/Sub with Cloud Run | - |
 
 For more Cloud Run samples beyond .NET, see the main list in the [Cloud Run Samples repository](https://github.com/GoogleCloudPlatform/cloud-run-samples).
 
@@ -105,3 +106,4 @@ for more information.
 [helloworld]: https://github.com/knative/docs/tree/master/docs/serving/samples/hello-world/helloworld-csharp
 [run_button_helloworld]: https://deploy.cloud.run/?git_repo=https://github.com/knative/docs&dir=docs/serving/samples/hello-world/helloworld-csharp
 [testing]: https://cloud.google.com/run/docs/testing/local#running_locally_using_docker_with_access_to_services
+[events_pubsub]: events-pubsub
