+++ run/events-pubsub/Dockerfile
@@ -23,5 +23,5 @@ WORKDIR /app
 COPY --from=build /app/out ./
 
 # Run the web service on container startup.
-ENTRYPOINT ["dotnet", "HelloEvents.dll"]
+ENTRYPOINT ["dotnet", "EventsPubSub.dll"]
