+++ run/events-pubsub/Program.cs
@@ -16,19 +16,22 @@ using System;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Hosting;
 
-namespace HelloEvents
+namespace EventsPubSub
 {
     public class Program
     {
+        // [START run_events_pubsub_server]
         public static void Main(string[] args)
         {
             CreateHostBuilder(args).Build().Run();
         }
+        // [END run_events_pubsub_server]
 
+        // [START run_events_pubsub_server_setup]
         public static IHostBuilder CreateHostBuilder(string[] args)
         {
-            string port = Environment.GetEnvironmentVariable("PORT") ?? "8080";
-            string url = String.Concat("http://0.0.0.0:", port);
+            var port = Environment.GetEnvironmentVariable("PORT") ?? "8080";
+            var url = $"http://0.0.0.0:{port}";
 
             return Host.CreateDefaultBuilder(args)
                 .ConfigureWebHostDefaults(webBuilder =>
@@ -36,5 +39,6 @@ namespace HelloEvents
                     webBuilder.UseStartup<Startup>().UseUrls(url);
                 });
         }
+        // [END run_events_pubsub_server_setup]
     }
 }
